% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStates.R
\name{createStates}
\alias{createStates}
\title{This counts the single states in a  data frame}
\usage{
createStates(data, s0, s1, s2, s3)
}
\arguments{
\item{data}{input your data set}

\item{s0}{input upper limit for state 0}

\item{s1}{input lower limit for state 1}

\item{s2}{input upper limit for state 2}

\item{s3}{input lower limit for state 3}
}
\value{
returns a small datafram with numbers of each state
}
\description{
This counts the single states in a  data frame
}
