% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagMat.R
\name{diagMat}
\alias{diagMat}
\title{Creates a transition matrix diagram}
\usage{
diagMat(matrix1)
}
\arguments{
\item{matrix1}{input matrix output-ed from transcount function}
}
\value{
diagram
}
\description{
Creates a transition matrix diagram
}
