% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyMeanF.R
\name{easyMeanF}
\alias{easyMeanF}
\title{Easy mean forecasting with barely any code}
\usage{
easyMeanF(data1, Y1, Y2)
}
\arguments{
\item{data1}{input a numerical vector that you want to forecast}

\item{Y1}{input the first year}

\item{Y2}{input the last year}
}
\value{
forecast for next year
}
\description{
Easy mean forecasting with barely any code
}
