% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcorr.R
\name{plotcorr}
\alias{plotcorr}
\title{Function that creates a scatter plot with a smoother, that allows users to select dots and see their values using shiny.}
\usage{
plotcorr(data, x, y, x1, x2)
}
\arguments{
\item{data}{input the data set that is to be used}

\item{x}{input the variable}

\item{y}{second input the variable}

\item{x1}{input the variable, but add data$x1}

\item{x2}{second input the variable, but add data$x2}
}
\value{
a shiny application that allows users to select dots in scatter plot and see data
}
\description{
Function that creates a scatter plot with a smoother, that allows users to select dots and see their values using shiny.
}
