% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transCount.R
\name{transcount}
\alias{transcount}
\title{Computes a transition matrix}
\usage{
transcount(matrix2, state0, state1, state2, state3)
}
\arguments{
\item{matrix2}{input data set}

\item{state0}{upper limit for state 0}

\item{state1}{lower limit for state 1}

\item{state2}{upper limit for state 2}

\item{state3}{lower limit for state 3}
}
\value{
A transition matrix
}
\description{
Computes a transition matrix
}
